pyproject.toml
[tool.poetry]
name = "agent-system"
version = "0.1.0"
description = "Agent System for KMMLU evaluation"
authors = ["Daehyun"]
license = "MIT"
readme = "README.md"
packages = [{include = "agent_system"}]  # 프로젝트 패키지 포함

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
openai = ">=1.61.1,<2.0.0"
pandas = ">=2.2.3,<3.0.0"
pytest = ">=8.3.4,<9.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
numpy = "^2.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.25.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"